my-dolphin-tool-02/requirements.txt
# ============================================
# BACKEND DEPENDENCIES (Python 3.10+)
# ============================================

# Flask Web Framework
Flask==3.0.3
Werkzeug==3.0.2
itsdangerous==2.2.0
blinker==1.8.2

# Flask Extensions & Authentication
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-JWT-Extended==4.6.0
WTForms==3.1.2

# Database - Use psycopg2-binary for compatibility with existing code
psycopg2-binary==2.9.9

# HTTP/API & Requests
requests==2.32.3
flask-cors==4.0.1
urllib3==2.1.0

# Production Server
gunicorn==22.0.0

# Environment & Config
python-dotenv==1.0.1

# Security & Encryption
cryptography==42.0.8
email-validator==2.2.0
PyJWT==2.11.0
bcrypt==4.1.2

# Background Job Processing (Optional)
redis==5.0.7
celery==5.4.0

# Hardware Communication
pyserial==3.5

# Utilities
Pillow==11.0.0
click==8.1.7

pycryptodome==3.20.0

flask-limiter>=3.5.0